pyproject.toml
[project]
name = "langchain-study"
version = "0.1.0"
description = "A project for studying LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.7",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain>=1.2.7",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]